waveshare213v2: Move error handler to separate file
The structure definition and the functions were already split up. Moving
to a separate file simplifies the code structure.
