+++ waveshare2in13v2/waveshare213v2.go
@@ -81,12 +81,6 @@ type Opts struct {
 // PartialUpdate defines if the display should do a full update or just a partial update.
 type PartialUpdate bool
 
-// errorHandler is a wrapper for error management.
-type errorHandler struct {
-	d   Dev
-	err error
-}
-
 const (
 	// Full should update the complete display.
 	Full PartialUpdate = false
@@ -134,48 +128,6 @@ var EPD2in13v2 = Opts{
 // display.
 func dataDimensions(opts *Opts) (int, int) {
 	return opts.Height, (opts.Width + 7) / 8
-}
-
-func (eh *errorHandler) rstOut(l gpio.Level) {
-	if eh.err != nil {
-		return
-	}
-	eh.err = eh.d.rst.Out(l)
-}
-
-func (eh *errorHandler) cTx(w []byte, r []byte) {
-	if eh.err != nil {
-		return
-	}
-	eh.err = eh.d.c.Tx(w, r)
-}
-
-func (eh *errorHandler) dcOut(l gpio.Level) {
-	if eh.err != nil {
-		return
-	}
-	eh.err = eh.d.dc.Out(l)
-}
-
-func (eh *errorHandler) csOut(l gpio.Level) {
-	if eh.err != nil {
-		return
-	}
-	eh.err = eh.d.cs.Out(l)
-}
-
-func (eh *errorHandler) sendCommand(cmd byte) {
-	if eh.err != nil {
-		return
-	}
-	eh.err = eh.d.sendCommand(cmd)
-}
-
-func (eh *errorHandler) sendData(d []byte) {
-	if eh.err != nil {
-		return
-	}
-	eh.err = eh.d.sendData(d)
 }
 
 // New creates new handler which is used to access the display.
